Change structure of common interfaces
