+++ src/PresentationVerifier.ts
@@ -1,0 +1,7 @@
+import * as _ from 'lodash';
+import { Credential } from './Credential';
+
+export type VerifyAnchorFunction = (credential : Credential) => boolean;
+
+export class PresentationVerifier {
+}
