+++ src/VerifiablePresentationManager.ts
@@ -1,4 +1,11 @@
 import * as _ from 'lodash';
+import {
+    ClaimIdentifier,
+    CredentialIdentifier,
+    CredentialProofLeave,
+    Credential
+} from './Credential';
+import { VerifyAnchorFunction } from './PresentationVerifier';
 
 /**
  * Used to setup VerifiablePresentationManager global behavior
@@ -31,15 +38,6 @@ export interface VPMOptions {
 }
 
 /**
- * A VerifiableCredential Identifier
- * This define the type of the verifiable credential or verifiable presentation
- */
-// TODO complete the list
-export type CredentialIdentifier = 'credential-cvc:Email-v1'
-    | 'credential-cvc:PhoneNumber-v1'
-    | 'credential-cvc:GenericDocumentId-v1';
-
-/**
  * An unique reference to a managed presentation
  */
 export interface PresentationReference {
@@ -52,14 +50,6 @@ export interface PresentationReference {
      */
     uid: string;
 }
-
-/**
- * A Verifiable Claim Identifier
- */
-// TODO complete the list
-export type ClaimIdentifier = 'credential-cvc:Email-v1'
-    | 'claim-cvc:Contact.phoneNumber-v1';
-
 /**
  * An unique reference to a managed claim
  */
@@ -98,56 +88,6 @@ export interface SearchClaimCriteria {
  * but can be linked to a claim. Making it possible to verify if that data was the same used to issue
  * the credential. This is useful for document images, selfies, etc...
  */
-
-/**
- * Credential Proof Leave Representation
- */
-export interface CredentialProofLeave {
-    /**
-     * see [[ClaimIdentifier]]
-     */
-    identifier: ClaimIdentifier;
-    /**
-     * The leave value
-     */
-    value: string;
-    /**
-     * The claim path
-     */
-    claimPath: string;
-}
-
-/**
- * Credential Proof Representation
- */
-export interface CredentialProof {
-    /**
-     * A list of [[CredentialProofLeave]]
-     */
-    leaves: CredentialProofLeave[];
-}
-
-/**
- * Credential representation
- */
-export interface Credential {
-    /**
-     * Unique identifier
-     */
-    id: string;
-    /**
-     * see [[CredentialIdentifier]]
-     */
-    identifier: CredentialIdentifier;
-    /**
-     * see [[CredentialProof]]
-     */
-    proof: CredentialProof;
-    /**
-     * Claim values (dynamic object)
-     */
-    claim: any;
-}
 
 /**
  * Evidence representation
@@ -205,7 +145,6 @@ export interface VerifiablePresentationManagerStatus {
 export type DSRJSON = string;
 
 
-export type VerifyFunction = (credential : Credential) => boolean;
 
 /**
  * Abstract all complexity about the Verifiable Credentials handling by providing utility methods
@@ -220,13 +159,13 @@ export class VerifiablePresentationManager {
     presentations: PresentationReference[];
     claims: AvailableClaim[];
     status: VerifiablePresentationManagerStatus;
-    verifyAnchor : VerifyFunction;
+    verifyAnchor : VerifyAnchorFunction;
 
     /**
      * @param options - Defines the global behavior and security of VerifiablePresentationManager
      * @param verifyAnchor - An async function that is able to verify the presentation anchor in a public Blockchain
      */
-    constructor(options: VPMOptions, verifyAnchor? : VerifyFunction) {
+    constructor(options: VPMOptions, verifyAnchor? : VerifyAnchorFunction) {
         this.options = options;
         this.presentations = [];
         this.claims = [];
