+++ src/__tests__/PresentationVerifier.test.ts
@@ -1,0 +1,8 @@
+import { PresentationVerifier } from '../PresentationVerifier';
+
+describe('PresentationVerifier', () => {
+    it('should initialize a PresentationVerifier', () => {
+        const presentationManager = new PresentationVerifier();
+        expect(presentationManager).toBeDefined();
+    });
+});
